move prepare config to components
